+++ packages/api/src/client/types/chat/bsky/convo/listConvos.ts
@@ -15,6 +15,8 @@ const id = 'chat.bsky.convo.listConvos'
 export interface QueryParams {
   limit?: number
   cursor?: string
+  readState?: 'unread' | (string & {})
+  status?: 'request' | 'accepted' | (string & {})
 }
 
 export type InputSchema = undefined