+++ packages/api/src/client/types/com/atproto/repo/applyWrites.ts
@@ -64,6 +64,7 @@ export function toKnownErr(e: any) {
 export interface Create {
   $type?: 'com.atproto.repo.applyWrites#create'
   collection: string
+  /** NOTE: maxLength is redundant with record-key format. Keeping it temporarily to ensure backwards compatibility. */
   rkey?: string
   value: { [_ in string]: unknown }
 }