+++ packages/api/src/client/types/com/atproto/sync/listRepos.ts
@@ -46,7 +46,14 @@ export interface Repo {
   rev: string
   active?: boolean
   /** If active=false, this optional field indicates a possible reason for why the account is not active. If active=false and no status is supplied, then the host makes no claim for why the repository is no longer being hosted. */
-  status?: 'takendown' | 'suspended' | 'deactivated' | (string & {})
+  status?:
+    | 'takendown'
+    | 'suspended'
+    | 'deleted'
+    | 'deactivated'
+    | 'desynchronized'
+    | 'throttled'
+    | (string & {})
 }
 
 const hashRepo = 'repo'