+++ packages/pds/src/api/com/atproto/repo/applyWrites.ts
@@ -130,7 +130,7 @@ export default function (server: Server, ctx: AppContext) {
         }
       })
 
-      await ctx.sequencer.sequenceCommit(did, commit, writes)
+      await ctx.sequencer.sequenceCommit(did, commit)
       await ctx.accountManager.updateRepoRoot(did, commit.cid, commit.rev)
 
       return {