+++ packages/pds/src/api/com/atproto/server/activateAccount.ts
@@ -42,17 +42,20 @@ export default function (server: Server, ctx: AppContext) {
           newBlocks: blocks.blocks,
           relevantBlocks: blocks.blocks,
           removedCids: new CidSet(),
+          ops: [],
+          blobs: new CidSet(),
+          prevData: null,
         }
       })
 
       // @NOTE: we're over-emitting for now for backwards compatibility, can reduce this in the future
       const status = await ctx.accountManager.getAccountStatus(requester)
       await ctx.sequencer.sequenceAccountEvt(requester, status)
-      await ctx.sequencer.sequenceHandleUpdate(
+      await ctx.sequencer.sequenceIdentityEvt(
         requester,
         account.handle ?? INVALID_HANDLE,
       )
-      await ctx.sequencer.sequenceCommit(requester, commitData, [])
+      await ctx.sequencer.sequenceCommit(requester, commitData)
     },
   })
 }