+++ packages/pds/src/api/com/atproto/server/createAccount.ts
@@ -74,7 +74,7 @@ export default function (server: Server, ctx: AppContext) {
         if (!deactivated) {
           await ctx.sequencer.sequenceIdentityEvt(did, handle)
           await ctx.sequencer.sequenceAccountEvt(did, AccountStatus.Active)
-          await ctx.sequencer.sequenceCommit(did, commit, [])
+          await ctx.sequencer.sequenceCommit(did, commit)
         }
         await ctx.accountManager.updateRepoRoot(did, commit.cid, commit.rev)
         await ctx.actorStore.clearReservedKeypair(signingKey.did(), did)