+++ packages/pds/src/index.ts
@@ -37,7 +37,7 @@ export { createSecretKeyObject } from './auth-verifier'
 export { type Handler as SkeletonHandler } from './lexicon/types/app/bsky/feed/getFeedSkeleton'
 export { createServer as createLexiconServer } from './lexicon'
 export * as sequencer from './sequencer'
-export { type PreparedWrite } from './repo'
+export { type CommitDataWithOps, type PreparedWrite } from './repo'
 export * as repoPrepare from './repo/prepare'
 export { scripts } from './scripts'